Do not require loader to be specified.
I need to be able to generate multiple ES6 output files, that will be
interoperable (think Ember CLI addons), but need to be able to share a
single loader.  The solution is quite simple (just have the loader in a
shared `vendor.js` that is loaded before the application code), but
requires that I provide an empty file to the `loaderFile` option here.